+++ src/lib/Text/index.js
